Page Container Component Pattern
