+++ src/pages/Configuration/index.js
@@ -1,0 +1,3 @@
+import ConfigurationManagementContainer from "../../containers/ConfigurationManagementContainer";
+
+export default ConfigurationManagementContainer;
