+++ src/pages/Dashboard/index.js
@@ -1,0 +1,3 @@
+import DashboardContainer from "../../containers/DashboardContainer";
+
+export default DashboardContainer;
