+++ src/pages/Monitor/index.js
@@ -1,0 +1,3 @@
+import MonitoringSystemContainer from "../../containers/MonitoringSystemContainer";
+
+export default MonitoringSystemContainer;
