+++ src/pages/Runtime/index.js
@@ -1,0 +1,3 @@
+import RuntimeManagementContainer from "../../containers/RuntimeManagementContainer";
+
+export default RuntimeManagementContainer;
