--- src/reducers/rootReducer.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import configurationManagement from './configurationManagement';
-import monitoringSystem from './monitoringSystem';
-import runtimeManagement from './runtimeManagement';
-
-export default combineReducers({
-    configurationManagement,
-    monitoringSystem,
-    runtimeManagement
-})
