+++ src/redux/configureStore.js
@@ -1,10 +1,10 @@
 import { createStore, applyMiddleware } from "redux";
 import thunk from "redux-thunk";
-import allReducers from './reducers';
+import rootReducer from './reducers';
 
 const configureStore = () => {
   const store = createStore(
-    allReducers,
+    rootReducer,
     applyMiddleware(thunk)
   );
   return store;
