+++ src/redux/reducers/configurationManagement.js
