+++ src/redux/reducers/index.js
@@ -1,6 +1,12 @@
 import { combineReducers } from "redux";
 import auth from "./auth";
+import configurationManagement from './configurationManagement';
+import monitoringSystem from './monitoringSystem';
+import runtimeManagement from './runtimeManagement';
 
 export default combineReducers({
-    auth
+    auth,
+    configurationManagement,
+    monitoringSystem,
+    runtimeManagement
 });
