+++ src/redux/reducers/monitoringSystem.js
